+++ apps/q-cafe/web/src/pos1-sections/index.ts
@@ -3,6 +3,7 @@ export * from './Pos1HeaderSection';
 export * from './Pos1ProductSection';
 export * from './Pos1BillingSection';
 export * from './Pos1BillsDrawer';
+export * from './Pos1PreviousBillsDrawer';
 export * from './Pos1ManualEntrySection';
 export * from './TypeaheadDropUp';
 export * from './Pos1PaymentCollector';